after creating a level, one can now play the level (ONLY WORKS FOR IDE)
