+++ src/main/java/net/liquidpineapple/pang/screens/LevelScreen.java
@@ -1,33 +1,21 @@
 package net.liquidpineapple.pang.screens;
 
+import lombok.Getter;
+
 import net.liquidpineapple.pang.Application;
 import net.liquidpineapple.pang.gui.Board;
-import net.liquidpineapple.pang.gui.NumberToken;
 import net.liquidpineapple.pang.gui.ScoreSystem;
 import net.liquidpineapple.pang.gui.TimeSystem;
 import net.liquidpineapple.pang.objects.Ball;
-import net.liquidpineapple.pang.objects.BallMovement;
-import net.liquidpineapple.pang.objects.GameObject;
-import net.liquidpineapple.pang.objects.Player;
-import net.liquidpineapple.pang.objects.playerschemes.Player1;
-import net.liquidpineapple.pang.objects.playerschemes.Player2;
 
-import org.w3c.dom.Document;
-import org.w3c.dom.Element;
-import org.w3c.dom.Node;
-import org.w3c.dom.NodeList;
+import net.liquidpineapple.pang.objects.GameObject;
+
 import org.xml.sax.SAXException;
 
 import java.awt.Graphics2D;
 import java.awt.image.ImageObserver;
-import java.io.IOException;
-import java.io.InputStream;
 import java.util.ArrayList;
 import java.util.LinkedList;
-
-import javax.imageio.ImageIO;
-import javax.xml.parsers.DocumentBuilder;
-import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
 
 /**
@@ -38,141 +26,13 @@ import javax.xml.parsers.ParserConfigurationException;
  */
 public class LevelScreen extends Screen {
 
+  @Getter
   private LinkedList<GameObject> hudObjectList;
 
 
   public LevelScreen() {
     super();
     hudObjectList = new LinkedList<GameObject>();
-  }
-
-  /**
-   * Method that parses a XML-file into a level.
-   *
-   * @param xmlFile - path/filename of the XML-file that should be parsed.
-   * @return - returns a new level.
-   */
-  public static LevelScreen createFromXml(String xmlFile)
-      throws IOException, ParserConfigurationException, SAXException {
-
-    LevelScreen output = new LevelScreen();
-    if (createFileReader(xmlFile) != null) {
-      Document doc = createFileReader(xmlFile);
-      String backgroundPath = "/sprites/"
-          + doc.getElementsByTagName("background").item(0).getTextContent();
-      output.backgroundImage = ImageIO.read(LevelScreen.class.getResource(backgroundPath));
-
-      int levelDifficulty = Integer.parseInt(doc.getElementsByTagName("difficulty")
-          .item(0).getTextContent());
-      output.setDifficulty(levelDifficulty);
-
-      for (Ball ball : loadBalls(doc.getElementsByTagName("ball"))) {
-        output.objectList.add(ball);
-      }
-
-      for (Player player : loadPlayer(doc.getElementsByTagName("player"))) {
-        output.objectList.add(player);
-      }
-      loadTime(doc);
-      TimeSystem.resetTime(loadTime(doc));
-
-      for (NumberToken token : TimeSystem.getTimePlaces()) {
-        output.hudObjectList.add(token);
-      }
-
-    }
-    for (NumberToken token : ScoreSystem.getPlaces()) {
-      output.hudObjectList.add(token);
-    }
-    Application.lifeSystem.updateLifes();
-    output.hudObjectList.add(Application.lifeSystem);
-    return output;
-  }
-
-  /**
-   * Method that reads the balls from the XML-file and parses them into Ball objects.
-   *
-   * @param ballList - NodeList of the different balls to be parsed.
-   * @return - Arraylist with Ball objects.
-   */
-  public static ArrayList<Ball> loadBalls(NodeList ballList) {
-    ArrayList<Ball> ballArray = new ArrayList<>();
-
-    for (int i = 0; i < ballList.getLength(); i++) {
-      Node ballNode = ballList.item(i);
-      Element ballElement = (Element) ballNode;
-
-      int intXpos = Integer.parseInt(ballElement.getElementsByTagName("x")
-          .item(0).getTextContent());
-      int intYpos = Integer.parseInt(ballElement.getElementsByTagName("y")
-          .item(0).getTextContent());
-
-      int size = Integer.parseInt(ballElement.getElementsByTagName("size").item(0)
-          .getTextContent());
-
-      String direction = ballElement.getElementsByTagName("direction").item(0).getTextContent();
-      BallMovement ballMovement = BallMovement.valueOf(direction);
-      Ball ball = new Ball(intXpos, intYpos, ballMovement, size);
-      ballArray.add(ball);
-    }
-    return ballArray;
-  }
-
-  /**
-   * Method that reads the players from the XML-file and parses them into Player objects.
-   * Simular to the loadBalls method
-   *
-   * @param playerlist - NodeList of the players that should be parsed.
-   * @return - ArrayList of Player objects.
-   */
-  public static ArrayList<Player> loadPlayer(NodeList playerlist) {
-    ArrayList<Player> playerArray = new ArrayList<>();
-    for (int i = 0; i < playerlist.getLength(); i++) {
-      Node playernode = playerlist.item(i);
-      Element playerElement = (Element) playernode;
-
-      int intXpos = Integer.parseInt(playerElement.getElementsByTagName("x")
-          .item(0).getTextContent());
-      int intYpos = Integer.parseInt(playerElement.getElementsByTagName("y")
-          .item(0).getTextContent());
-
-      if (i == 0) {
-        playerArray.add(new Player(intXpos, intYpos, new Player1()));
-      } else if (i == 1 && Application.multiplayer) {
-        playerArray.add(new Player(intXpos, intYpos, new Player2()));
-      }
-    }
-    return playerArray;
-  }
-
-  /**
-   * Method to read time form an XML-file.
-   *
-   * @param doc {@link Document} to be read from
-   * @return - returns an int contained in the {@code <time></time>}
-   */
-  public static int loadTime(Document doc) {
-    return Integer.parseInt(doc.getElementsByTagName("time").item(0).getTextContent());
-  }
-
-  /**
-   * Method to create a FileReader to read the XML-file.
-   *
-   * @param xmlFile - Path/name of the XML-file to be parsed.
-   * @return - returns a new FileReader if no exception is thrown, else it will return null.
-   */
-  public static Document createFileReader(String xmlFile)
-      throws ParserConfigurationException, SAXException {
-
-    DocumentBuilder documentBuilder = DocumentBuilderFactory.newInstance().newDocumentBuilder();
-    InputStream in = LevelScreen.class.getResourceAsStream(xmlFile);
-    try {
-      return documentBuilder.parse(in);
-    } catch (IOException | IllegalArgumentException ex) {
-      //TODO Fix this shit
-      ex.printStackTrace();
-      return null;
-    }
   }
 
   @Override
@@ -206,7 +66,7 @@ public class LevelScreen extends Screen {
    */
   private void nextLevel() throws ParserConfigurationException, SAXException {
     Board currentBoard = Application.getBoard();
-    Screen newScreen = null;
+    Screen newScreen;
 
     if (currentBoard.getLevels().hasNext()) {
       newScreen = (Screen) currentBoard.getLevels().next();
