+++ src/main/java/net/liquidpineapple/pang/screens/UserCreatedLevels.java
@@ -1,0 +1,17 @@
+package net.liquidpineapple.pang.screens;
+
+/**
+ * Created by Tim on 13-10-2016.
+ * Class which keeps track of the userCreatedLevels.
+ * Part of the iterator design pattern.
+ */
+public class UserCreatedLevels extends LevelIterator {
+  private static final String DIRECTORY = "levels/customplayer";
+
+  /**
+   * Method to fill the levels array with userCreatedLevels.
+   */
+  public UserCreatedLevels() {
+    super.initArray(DIRECTORY);
+  }
+}
